BDOG-3888: restructure tests to separate model / connector tests
